pyproject.toml
[project]
name = "mlops-lab-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.3"
dependencies = [
    "argparse==1.4.0",
    "beautifulsoup4==4.14.3",
    "dvc==3.67.0",
    "hydra-core==1.3.2",
    "jupyter==1.1.1",
    "kagglehub==1.0.0",
    "matplotlib==3.10.8",
    "mlflow==3.10.0",
    "nltk==3.9.3",
    "numpy==2.4.2",
    "optuna==4.8.0",
    "pandas==2.3.3",
    "scikit-learn==1.8.0",
    "seaborn==0.13.2",
    "textblob==0.19.0",
    "wordcloud==1.9.6",
]
